Update geocode.R
updated api address and made small change to replace "&" with "and" in request. This would require the package stringr.  
Also, for some reason RCurl did not work for me. Replaced it with GET from httr. 
Sorry, this is the first time I propose changes. Please let me know if something is off. Thanks.